refactor: favorSetting 화면 분리
