+++ apps/enfpy-native/src/screens/favor-setting/FavorSetting.tsx
